Granny-voice/pyproject.toml
[project]
name = "granny-voice"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-deepgram>=0.6.0",
    "livekit-plugins-openai>=0.7.0",
    "livekit-plugins-elevenlabs>=0.2.0",
    "livekit-plugins-silero>=0.6.0",
    "livekit-api>=0.5.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "livekit-plugins-noise-cancellation~=0.2",
]

[tool.uv.workspace]
members = [
    "livekit-voice-agent",
    "livekit-voice-agent/livekit-voice-agent",
]

